+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 Fixed
 -----
 - Set ``value`` to ``AnnotationValue`` object on duplication when it is created
+- Send ``post_duplicate`` signal only on successful duplication
 
 
 ===================